Extract URLSession extension into its own file
It wasn't used in XcodeManager anymore and is large enough to warrant its own file.
